requirements.txt
canonicalwebteam.flask-base==0.7.3
canonicalwebteam.discourse==2.1.0
canonicalwebteam.search==0.2.1
